% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_price.R
\name{batch_price}
\alias{batch_price}
\title{Price of a Batch}
\usage{
batch_price(batch_id = "")
}
\arguments{
\item{batch_id}{ID for the batch}
}
\description{
The price for processing the batch.
}
\examples{
\dontrun{
batch_price("batch_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}

