% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_batch.R
\name{create_batch}
\alias{create_batch}
\title{Create Batch}
\usage{
create_batch(batch_name = "")
}
\arguments{
\item{file_path}{path to image file}
}
\description{
Create Batch
}
\examples{
\dontrun{
create_batch(batch_name="name_of_batch")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}

