% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_batch.R
\name{delete_batch}
\alias{delete_batch}
\title{Delete a Batch}
\usage{
delete_batch(batch_id = "")
}
\arguments{
\item{batch_id}{ID for the batch}
}
\description{
Delete a Batch
}
\examples{
\dontrun{
delete_batch("batch_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-batches}
}
