% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_batch_details.R
\name{get_batch_details}
\alias{get_batch_details}
\title{Get Details of a particular batch}
\usage{
get_batch_details(batch_id = "")
}
\arguments{
\item{batch_id}{ID for the batch}
}
\value{
list of length 26.
}
\description{
Get details such as who the batch was created by, user id of the creator, name of the batch, whether or not the batch has been submitted for processing, files in the batch, etc.
}
\examples{
\dontrun{
get_batch_details("batch_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-batch}
}

