% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_template_id.R
\name{get_template_id}
\alias{get_template_id}
\title{Set Template ID}
\usage{
get_template_id()
}
\description{
To digitize documents, create a template using the Captricity Web UI at \url{https://shreddr.captricity.com/job/}
The template tells Captricity which data to get from where in the document.
Get the template id of the template
}
\examples{
\dontrun{
get_template_id()
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}

