% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_batch_files.R
\name{list_batch_files}
\alias{list_batch_files}
\title{List all the files in a batch}
\usage{
list_batch_files(batch_id = "", ...)
}
\arguments{
\item{batch_id}{ID for the batch; Required; String}

\item{\dots}{Additional arguments passed to \code{\link{captr_GET}}.}
}
\value{
list. If no files in a batch, an empty list
}
\description{
List all the files in a batch
}
\examples{
\dontrun{
list_batch_files("batch_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}

