% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_batches.R
\name{list_batches}
\alias{list_batches}
\title{List Batches}
\usage{
list_batches(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{captr_GET}}.}
}
\value{
named list. Prints number of batches by default. Named list has 15 items:
\code{total_batch_cost_in_fields, user_included_fields_per_page, can_be_priced,
user_pay_go_fields_applied, page_count, total_user_cost_in_cents, user_subscription_fields_applied, 
user_subscription_fields_per_month, overage_field_count, total_user_cost_in_fields, batch_id,
user_subscription_fields, user_pay_go_fields, user_fields_per_overage_field, user_fields_per_page
}
}
\description{
A list of batches owned by the calling account.
}
\examples{
\dontrun{
list_batches()
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-batches}
}
