% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_docs.R
\name{list_docs}
\alias{list_docs}
\title{List Documents}
\usage{
list_docs()
}
\description{
A list of document resources owned by the calling account.
}
\examples{
\dontrun{
list_docs()
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-documents}
}

