% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list_instance_sets.R
\name{list_instance_sets}
\alias{list_instance_sets}
\title{List Instance Sets}
\usage{
list_instance_sets(job_id = "")
}
\arguments{
\item{job_id}{ID for the job (which you get from related_job_id field of submit_batch)}
}
\description{
List all the forms (instance sets) for a particular job.
Use this to get ids of forms (instance sets) you want to download.
}
\examples{
\dontrun{
list_instance_sets(job_id = "job_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}

