% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_jobs.R
\name{list_jobs}
\alias{list_jobs}
\title{List Jobs}
\usage{
list_jobs(...)
}
\description{
A list of jobs owned by the calling account.
}
\examples{
\dontrun{
list_jobs()
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-jobs}
}
