% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_batch_template.R
\name{set_batch_template}
\alias{set_batch_template}
\title{Assign Template ID to a Batch}
\usage{
set_batch_template(batch_id = "", template_id = "")
}
\arguments{
\item{batch_id}{Batch ID}

\item{template_id}{ID for the template}
}
\description{
To digitize documents, create a template using the Captricity Web UI at \url{https://shreddr.captricity.com/job/}
The template tells Captricity which data to get from where in the document.
Set the relevant document id using this function
}
\examples{
\dontrun{
set_template_id("batch_id", "template_id)
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}

