% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_template_id.R
\name{set_template_id}
\alias{set_template_id}
\title{Set Template ID}
\usage{
set_template_id(template_id = "")
}
\arguments{
\item{template_id}{ID for the template}
}
\description{
To digitize documents, create a template using the Captricity Web UI at \url{https://shreddr.captricity.com/job/}
The template tells Captricity which data to get from where in the document.
Set the relevant document id using this function
}
\examples{
\dontrun{
set_template_id(template_id="template_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}

