% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_token.R
\name{set_token}
\alias{set_token}
\title{Sets Application ID and Secret}
\usage{
set_token(appdetails = NULL)
}
\arguments{
\item{appdetails}{Required; A vector of client_id, client_secret. Get these from \url{https://shreddr.captricity.com/developer/}.
Set them before you use other functions.}
}
\description{
Set Client ID and Secret. Needed for interfacing with Clarifai. Run this before anything else.
}
\examples{
\dontrun{
set_token("app_token")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}
\keyword{Client}
\keyword{ID}
\keyword{Secret}
\keyword{Sets}
\keyword{and}

