% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{split_position}
\alias{split_position}
\title{Splits a UCSC position into three valid values}
\usage{
split_position(chr)
}
\arguments{
\item{x}{A character object in the format chr:start-stop}
}
\value{
A 3-column data frame form the values in \code{x}.
}
\description{
Splits a UCSC position into three valid values
}
\examples{
split_position

}
