% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{strip_comma}
\alias{strip_comma}
\title{Converts comma-containing strings to numeric values}
\usage{
strip_comma(x)
}
\arguments{
\item{x}{A character object}
}
\value{
The numeric value of \code{x}.
}
\description{
Converts comma-containing strings to numeric values
}
\examples{
strip_comma(c("1,000", "2,100"))
}
