% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_batch.R
\name{submit_batch}
\alias{submit_batch}
\title{Submit the Batch for Processing}
\usage{
submit_batch(batch_id = "", ...)
}
\arguments{
\item{batch_id}{ID for the batch. Required. String.}

\item{\dots}{Additional arguments passed to \code{\link{captr_POST}}.}
}
\description{
Submit the Batch for Processing
}
\examples{
\dontrun{
submit_batch("batch_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}
