% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_readiness.R
\name{test_readiness}
\alias{test_readiness}
\title{Test Readiness of a Batch}
\usage{
test_readiness(batch_id = "", ...)
}
\arguments{
\item{batch_id}{ID for the batch. Required. String.}

\item{\dots}{Additional arguments passed to \code{\link{captr_GET}}.}
}
\description{
Check if the batch is ready to be processed
}
\examples{
\dontrun{
test_readiness("batch_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}
