% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_progress.R
\name{track_progress}
\alias{track_progress}
\title{Track Progress of a Job}
\usage{
track_progress(job_id = "")
}
\arguments{
\item{job_id}{ID for the job (which you get from related_job_id field of submit_batch)}
}
\description{
Check what percentage of the job is digitized
}
\examples{
\dontrun{
track_progress("job_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}
