% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_profile.R
\name{user_profile}
\alias{user_profile}
\title{User Profile}
\usage{
user_profile(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{captr_GET}}.}
}
\value{
named list with the following elements: \code{username, deny_batch_submission_for_digitization,
account_balance, user_id, enrolled_subscription_is_annualized, enrolled_subscription, 
enrolled_subscription_is_trial, trial_eligible, remaining_quota}
}
\description{
Provides account information for the provided API Token.
}
\examples{
\dontrun{
user_profile()
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-user_profile}
}
