% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/captr.R
\docType{package}
\name{captr-package}
\alias{captr}
\alias{captr-package}
\title{captr: R Client for the Capticrity API}
\description{
OCR images and forms. See \url{https://captricity.com/}.
Details about results of calls to the API can be found at \url{https://shreddr.captricity.com/developer/}.

Your need credentials to use this application.
If you haven't already, you can get this at \url{https://shreddr.captricity.com/developer/}.
}
\author{
Gaurav Sood
}

