% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_batch.R
\name{create_batch}
\alias{create_batch}
\title{Create Batch}
\usage{
create_batch(batch_name = "")
}
\arguments{
\item{batch_name}{name of the batch; Required; character}
}
\value{
List of length 26. Includes information like created_by, user_id, etc.
}
\description{
Create a new batch.
}
\examples{
\dontrun{
create_batch(batch_name="name_of_batch")
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-batch}
}

