% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_job.R
\name{delete_job}
\alias{delete_job}
\title{Delete a Job}
\usage{
delete_job(job_id = "", ...)
}
\arguments{
\item{job_id}{ID for the job (which you get from related_job_id field of submit_batch)}

\item{\dots}{Additional arguments passed to \code{\link{captr_DELETE}}.}
}
\description{
Delete a Job
}
\examples{
\dontrun{
delete_job("job_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-batches}
}
