% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all.R
\name{get_all}
\alias{get_all}
\title{Get All the Data For a Particular Job in a csv}
\usage{
get_all(job_id = "", output_dir = "./")
}
\arguments{
\item{job_id}{ID for the job}

\item{output_dir}{output directory}
}
\description{
Get all the data
}
\examples{
\dontrun{
get_all(job_id ="job_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}
