% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_template_id.R
\name{get_template_id}
\alias{get_template_id}
\title{Set Template ID}
\usage{
get_template_id(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{captr_GET}}.}
}
\value{
named list
}
\description{
To digitize documents, create a template using the Captricity Web UI at \url{https://shreddr.captricity.com/job/} 
The template tells Captricity which data to get from where in the document.
}
\details{
Get the template ids of all the documents
}
\examples{
\dontrun{
get_template_id()
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}
