% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_batch_files.R
\name{list_batch_files}
\alias{list_batch_files}
\title{Test Readiness of a Batch}
\usage{
list_batch_files(batch_id = "")
}
\arguments{
\item{batch_id}{ID for the batch}
}
\description{
Check if the batch is ready to be processed
}
\examples{
\dontrun{
list_batch_files("batch_id")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}

