% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_batches.R
\name{list_batches}
\alias{list_batches}
\title{List Batches}
\usage{
list_batches()
}
\description{
A list of batches owned by the calling account.
}
\examples{
\dontrun{
list_batches()
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-batches}
}

