% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_batches.R, R/user_profile.R
\name{list_batches}
\alias{list_batches}
\title{List Batches}
\usage{
list_batches()

list_batches()
}
\description{
A list of batches owned by the calling account.

Provides account information for the provided API Token.
}
\examples{
\dontrun{
list_batches()
}
\dontrun{
list_batches()
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-batches}

\url{https://shreddr.captricity.com/developer/api-reference/#v1-user_profile}
}

