% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_token.R
\name{set_token}
\alias{set_token}
\title{Sets Application Token}
\usage{
set_token(app_token = NULL, force = FALSE)
}
\arguments{
\item{app_token}{Application token. Get these from \url{https://shreddr.captricity.com/developer/}.}

\item{force}{Force change the \code{CaptricityToken} stored in the environment}
}
\description{
Captricity requires an application token to use the API. Get the token from \url{https://shreddr.captricity.com/}.
The functions looks for \code{CaptricityToken} in the environment. If it doesn't find it or if change is forced, 
it looks for arguments passed in the function. If it fails to find that, it asks for input.
}
\details{
Run this function before anything else.
}
\examples{
\dontrun{
set_token("app_token")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}
\keyword{Application}
\keyword{Sets}
\keyword{Token}
