% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_image.R
\name{upload_image}
\alias{upload_image}
\title{Upload Image}
\usage{
upload_image(batch_id = "", path_to_image = "")
}
\arguments{
\item{batch_id}{ID for the batch}

\item{path_to_image}{Path to the image you want OCRd}
}
\description{
Upload an image to captricity.
}
\examples{
\dontrun{
upload_image("batch_id", "path_to_image")
}
}
\references{
\url{https://shreddr.captricity.com/developer/}
}

