% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_profile.R
\name{user_profile}
\alias{user_profile}
\title{User Profile}
\usage{
user_profile(...)
}
\arguments{
\item{\dots}{Additional arguments passed to \code{\link{captr_GET}}.}
}
\value{
named list
}
\description{
Provides account information for the provided API Token.
}
\examples{
\dontrun{
user_profile()
}
}
\references{
\url{https://shreddr.captricity.com/developer/api-reference/#v1-user_profile}
}

